GEODE-6534: Extract NonTXEntry from LocalRegion
