+++ geode-core/src/integrationTest/java/org/apache/geode/cache/query/functional/QueryREUpdateInProgressJUnitTest.java
@@ -37,7 +37,7 @@ import org.apache.geode.cache.query.SelectResults;
 import org.apache.geode.cache.query.data.Portfolio;
 import org.apache.geode.cache.query.types.ObjectType;
 import org.apache.geode.internal.cache.GemFireCacheImpl;
-import org.apache.geode.internal.cache.LocalRegion.NonTXEntry;
+import org.apache.geode.internal.cache.NonTXEntry;
 import org.apache.geode.internal.cache.RegionEntry;
 import org.apache.geode.test.junit.categories.OQLIndexTest;
