+++ geode-core/src/main/java/org/apache/geode/cache/query/internal/index/DummyQRegion.java
@@ -30,6 +30,7 @@ import org.apache.geode.cache.query.types.ObjectType;
 import org.apache.geode.internal.cache.CachedDeserializable;
 import org.apache.geode.internal.cache.HasCachePerfStats;
 import org.apache.geode.internal.cache.LocalRegion;
+import org.apache.geode.internal.cache.NonTXEntry;
 import org.apache.geode.internal.cache.RegionEntry;
 import org.apache.geode.internal.offheap.StoredObject;
 import org.apache.geode.internal.offheap.annotations.Released;
@@ -196,7 +197,7 @@ public class DummyQRegion extends QRegion {
     while (!(rgn instanceof LocalRegion)) {
       rgn = ((QRegion) TypeUtils.checkCast(rgn, QRegion.class)).getRegion();
     }
-    entries.add(((LocalRegion) rgn).new NonTXEntry(entry));
+    entries.add(new NonTXEntry((LocalRegion) rgn, entry));
     return entries;
   }
 
@@ -212,7 +213,7 @@ public class DummyQRegion extends QRegion {
 
   @Override
   public Region.Entry getEntry(Object key) {
-    LocalRegion.NonTXEntry e = (LocalRegion.NonTXEntry) super.getEntry(key);
+    NonTXEntry e = (NonTXEntry) super.getEntry(key);
     Region.Entry retVal = null;
     if (e != null && this.entry == e.getRegionEntry()) {
       retVal = e;
