+++ geode-core/src/main/java/org/apache/geode/internal/cache/EntriesSet.java
@@ -25,7 +25,6 @@ import java.util.NoSuchElementException;
 import org.apache.geode.cache.EntryDestroyedException;
 import org.apache.geode.cache.Region;
 import org.apache.geode.internal.cache.LocalRegion.IteratorType;
-import org.apache.geode.internal.cache.LocalRegion.NonTXEntry;
 import org.apache.geode.internal.cache.entries.AbstractRegionEntry;
 
 /** Set view of entries */
